add admin directory in /pages move all admin pages in /pages/admin
