+++ src/frontend/webappService/src/pages/admin/AddOrder.js
@@ -1,8 +1,8 @@
 import React from 'react'
 import '../css/App.css'
-import Navbar from '../components/Navbar.js'
-import OrderForm from '../components/OrderForm'
-import CheckUserConnection from '../components/CheckUserConnection'
+import Navbar from '../../components/Navbar.js'
+import OrderForm from '../../components/OrderForm.js'
+import CheckUserConnection from '../../components/CheckUserConnection.js'
 
 function AddOrder() {
   return (
