+++ src/frontend/webappService/src/pages/admin/AddRole.js
@@ -1,7 +1,7 @@
 import React from 'react';
-import RoleForm from '../components/RoleForm';
-import Navbar from '../components/Navbar';
-import CheckAdminConnection from '../components/CheckAdminConnection';
+import RoleForm from '../../components/RoleForm';
+import Navbar from '../../components/Navbar';
+import CheckAdminConnection from '../../components/CheckAdminConnection';
 
 function AddRole() {
   return (
